// limesz_client/projects/pluto/src/lib/api-providers/generated-api/models/resource-prop.ts
/* tslint:disable */
/* eslint-disable */
import { ResourcePropOptions } from '../models/resource-prop-options';
export interface ResourceProp {
  embededTypeDefinition?: any | null;
  jsAccessor?: string | null;
  propName?: string | null;
  propOptions?: ResourcePropOptions;
  propType?: string | null;
}